Initial commit
A set of Rust bindings for the Xapian search engine

This crate contains a mixture of generated bindings (via `autocxx`) and
manual ones (via `cxx`), as well as higher-level, more idiomatic Rust
wrappers for the same.

This commit brings support for basic querying and indexing, including
iteration of search results and document terms.