Add support for implementing Stoppers in Rust
This commit introduces a number of shims and wrapper types to enable
the user to implement custom Xapian::Stopper behavior by implementing a
Rust trait. In this implementation, the trait is leaked before it is
handed off to Xapian, so that it survives for the duration of the
program. This is (hopefully) a short-term workaround.

A custom C++ class is introduced since `autocxx` does not currently
handle `operator()`. The C++ class redirects to a method `autocxx` can
handle correctly, and the Rust implementation defers to a
dynamically-dispatched instance of the `Stopper` trait.