Expose `RangeProcessor` as a trait
This **breaking** change removes the NativeRangeProcessor machinery, in
favor of a RangeProcessor trait, along with concrete and blanket
implementations of it.

The following concrete implementations are provided:
- `DateRangeProcessor`
- `DateTimeRangeProcessor`
- `NumberRangeProcessor`
- `StringRangeProcessor`

Likewise, a blanket implementation is provided so that a
function/closure matching `Fn(&str, &str) -> (Option<T>, Option<T>)`
can be used as a `RangeProcessor` as long as `T` implements `ToValue`.